MAINT: Rename optimize module and make them private
