+++ scipy/optimize/_linesearch.py
