+++ scipy/optimize/_slsqp_py.py
@@ -20,7 +20,7 @@ import numpy as np
 from scipy.optimize._slsqp import slsqp
 from numpy import (zeros, array, linalg, append, asfarray, concatenate, finfo,
                    sqrt, vstack, exp, inf, isfinite, atleast_1d)
-from .optimize import (OptimizeResult, _check_unknown_options,
+from ._optimize import (OptimizeResult, _check_unknown_options,
                        _prepare_scalar_function, _clip_x_for_func,
                        _check_clip_x)
 from ._numdiff import approx_derivative
