+++ scipy/optimize/_tnc.py
@@ -33,7 +33,7 @@ value of the function, and whose second argument is the gradient of the function
 """
 
 from scipy.optimize import moduleTNC
-from .optimize import (MemoizeJac, OptimizeResult, _check_unknown_options,
+from ._optimize import (MemoizeJac, OptimizeResult, _check_unknown_options,
                        _prepare_scalar_function)
 from ._constraints import old_bound_to_new
