+++ scipy/optimize/_zeros_py.py
